fix(pr6-followup): cast ff_player_id to long, not int (Spark ANSI CAST_OVERFLOW, second site)
Re-trigger of compute_defcon_lite (run_id=699730105853777) after PR #209
landed STILL failed with [CAST_OVERFLOW] LONG→INTEGER. Wheel inspection
confirmed 0.3.15 contained PR #209's ff_team_id `.cast("long")` fix —
so the offending cast was elsewhere.

Found it: src/ingestion/defcon_lite_360.py:247 (pre-PR-6-followup code)
casts the synthetic ff_player_id to int:

  ff_df.withColumn("ff_player_id", F.monotonically_increasing_id().cast("int"))

`monotonically_increasing_id()` returns 64-bit values (upper 31 bits =
partition id, lower 33 bits = record number within partition). Under
multi-partition serverless the values exceed 2^31 and Spark ANSI mode
rejects LONG→INT statically. The cast had been dormant since the original
DEFCON ingestion: SkipGuard prevented re-execution against the bronze
table after the initial populate, so this static-type rejection never
surfaced. PR-6-followup's DELETE FROM bronze.defcon_results forced
re-ingest and exposed the latent bug.

Fix: change .cast("int") → .cast("long"). Matches credits_schema
StructField("defender_player_id", LongType()) which receives this column
downstream. Paired with PR #209's ff_team_id fix, defcon_lite_360.py is
now ANSI-clean.

Pre-push verified: ruff clean; pytest test_trigger_dbt_job 9/9 pass.

+++ src/ingestion/defcon_lite_360.py
@@ -243,8 +243,13 @@ def process_360_matches(
 
     # 360 freeze frames are anonymous — synthesize player_id; team_id
     # is derived post-join below from the teammate boolean.
+    # ff_player_id MUST be cast to long, not int: monotonically_increasing_id()
+    # returns 64-bit values (upper 31 bits = partition id, lower 33 bits = row
+    # number within partition); under multi-partition serverless the values
+    # exceed 2^31 and Spark ANSI mode rejects LONG→INT statically. Matches
+    # credits_schema StructField("defender_player_id", LongType()) downstream.
     ff_df = (
-        ff_df.withColumn("ff_player_id", F.monotonically_increasing_id().cast("int"))
+        ff_df.withColumn("ff_player_id", F.monotonically_increasing_id().cast("long"))
         .withColumn("ff_velocity_x", F.lit(0.0))
         .withColumn("ff_velocity_y", F.lit(0.0))
     )
